Close #367 - Move ToFuture to core and leave only typeclass instances in the sub-projects
