+++ cats-effect/src/test/scala-2/effectie/cats/toFutureSpec.scala
@@ -2,6 +2,8 @@ package effectie.cats
 
 import cats._
 import cats.effect._
+import effectie.cats.toFuture._
+import effectie.core.ToFuture
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
 import hedgehog._
@@ -14,7 +16,7 @@ import scala.concurrent.duration._
 /** @author Kevin Lee
   * @since 2020-09-23
   */
-object ToFutureSpec extends Properties {
+object toFutureSpec extends Properties {
   private implicit val errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
 
   override def tests: List[Test] = List(
@@ -23,18 +25,14 @@ object ToFutureSpec extends Properties {
       IoSpec.testUnsafeToFuture
     ),
     property(
-      "test ToFuture[Future].unsafeToFuture",
-      FutureSpec.testUnsafeToFuture
-    ),
-    property(
       "test ToFuture[Id].unsafeToFuture",
       IdSpec.testUnsafeToFuture
     )
   )
 
-  private val waitFor800Millis = WaitFor(800.milliseconds)
-
   object IoSpec {
+    private val waitFor800Millis = WaitFor(800.milliseconds)
+
     @SuppressWarnings(Array("org.wartremover.warts.IsInstanceOf"))
     def testUnsafeToFuture: Property = for {
       a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
@@ -65,39 +63,9 @@ object ToFutureSpec extends Properties {
 
   }
 
-  private val waitFor300Millis = WaitFor(300.milliseconds)
-
-  object FutureSpec {
-    @SuppressWarnings(Array("org.wartremover.warts.IsInstanceOf"))
-    def testUnsafeToFuture: Property = for {
-      a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
-    } yield {
-      implicit val es: ExecutorService = ConcurrentSupport.newExecutorService(2)
-      @SuppressWarnings(Array("org.wartremover.warts.ExplicitImplicitTypes"))
-      implicit val ec                  =
-        ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
-      ConcurrentSupport.runAndShutdown(es, waitFor300Millis) {
-        val expected = Future(a)
-        val fa       = Future(a)
-
-        val future = ToFuture[Future].unsafeToFuture(fa)
-        val actual = ConcurrentSupport.futureToValueAndTerminate(es, waitFor300Millis)(future)
-
-        Result.all(
-          List(
-            Result
-              .assert(future.isInstanceOf[Future[Int]])
-              .log(s"future is not an instance of Future[Int]. future.getClass: ${future.getClass.toString}"),
-            actual ==== ConcurrentSupport.futureToValueAndTerminate(es, waitFor300Millis)(expected),
-            actual ==== a
-          )
-        )
-      }
-    }
-
-  }
-
   object IdSpec {
+    private val waitFor300Millis = WaitFor(300.milliseconds)
+
     @SuppressWarnings(Array("org.wartremover.warts.IsInstanceOf"))
     def testUnsafeToFuture: Property = for {
       a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
