+++ cats-effect/src/test/scala-3/effectie/cats/toFutureSpec.scala
@@ -2,29 +2,27 @@ package effectie.cats
 
 import cats.*
 import cats.effect.*
+import effectie.cats.toFuture.given
+import effectie.core.ToFuture
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
 import hedgehog.*
 import hedgehog.runner.*
 
 import java.util.concurrent.ExecutorService
-import scala.concurrent.{ExecutionContext, Future}
 import scala.concurrent.duration.*
+import scala.concurrent.{ExecutionContext, Future}
 
 /** @author Kevin Lee
   * @since 2020-09-23
   */
-object ToFutureSpec extends Properties {
+object toFutureSpec extends Properties {
   private given errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
 
   override def tests: List[Test] = List(
     property(
       "test ToFuture[IO].unsafeToFuture",
       IoSpec.testUnsafeToFuture
-    ),
-    property(
-      "test ToFuture[Future].unsafeToFuture",
-      FutureSpec.testUnsafeToFuture
     ),
     property(
       "test ToFuture[Id].unsafeToFuture",
@@ -39,7 +37,8 @@ object ToFutureSpec extends Properties {
       val fa = IO(a)
 
       given es: ExecutorService  = ConcurrentSupport.newExecutorService(2)
-      given ec: ExecutionContext = ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
+      given ec: ExecutionContext =
+        ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
       ConcurrentSupport.runAndShutdown(es, WaitFor(800.milliseconds)) {
         val future   = ToFuture[IO].unsafeToFuture(fa)
         val expected = a
@@ -58,33 +57,6 @@ object ToFutureSpec extends Properties {
 
   }
 
-  object FutureSpec {
-    def testUnsafeToFuture: Property = for {
-      a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
-    } yield {
-      given es: ExecutorService  = ConcurrentSupport.newExecutorService(2)
-      given ec: ExecutionContext = ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
-      ConcurrentSupport.runAndShutdown(es, WaitFor(300.milliseconds)) {
-        val expected = Future(a)
-        val fa       = Future(a)
-
-        val future = ToFuture[Future].unsafeToFuture(fa)
-        val actual = ConcurrentSupport.futureToValueAndTerminate(es, WaitFor(300.milliseconds))(future)
-
-        Result.all(
-          List(
-            Result
-              .assert(future.isInstanceOf[Future[Int]])
-              .log(s"future is not an instance of Future[Int]. future.getClass: ${future.getClass.toString}"),
-            actual ==== ConcurrentSupport.futureToValueAndTerminate(es, WaitFor(300.milliseconds))(expected),
-            actual ==== a
-          )
-        )
-      }
-    }
-
-  }
-
   object IdSpec {
     def testUnsafeToFuture: Property = for {
       a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
@@ -92,7 +64,8 @@ object ToFutureSpec extends Properties {
       given es: ExecutorService = ConcurrentSupport.newExecutorService(2)
 
       val fa                     = a
-      given ec: ExecutionContext = ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
+      given ec: ExecutionContext =
+        ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
       ConcurrentSupport.runAndShutdown(es, WaitFor(300.milliseconds)) {
         val expected = Future(a)
