+++ cats-effect3/src/main/scala-2/effectie/cats/toFuture.scala
@@ -2,30 +2,18 @@ package effectie.cats
 
 import cats.Id
 import cats.effect._
+import effectie.core.ToFuture
 
 import scala.concurrent.{ExecutionContext, Future}
 
 /** @author Kevin Lee
   * @since 2020-09-23
   */
-trait ToFuture[F[_]] {
-
-  def unsafeToFuture[A](fa: F[A]): Future[A]
-
-}
-
-object ToFuture {
-
-  def apply[F[_]: ToFuture]: ToFuture[F] = implicitly[ToFuture[F]]
+object toFuture {
 
   implicit def ioToFuture(implicit runtime: unsafe.IORuntime): ToFuture[IO] = new ToFuture[IO] {
 
     override def unsafeToFuture[A](fa: IO[A]): Future[A] = fa.unsafeToFuture()
-  }
-
-  implicit val futureToFuture: ToFuture[Future] = new ToFuture[Future] {
-
-    override def unsafeToFuture[A](fa: Future[A]): Future[A] = fa
   }
 
   @SuppressWarnings(Array("org.wartremover.warts.ImplicitParameter"))
