+++ cats-effect3/src/main/scala-3/effectie/cats/toFuture.scala
@@ -2,30 +2,17 @@ package effectie.cats
 
 import cats.Id
 import cats.effect.*
+import effectie.core.ToFuture
 
 import scala.concurrent.{ExecutionContext, Future}
 
 /** @author Kevin Lee
   * @since 2020-09-23
   */
-trait ToFuture[F[_]] {
-
-  def unsafeToFuture[A](fa: F[A]): Future[A]
-
-}
-
-object ToFuture {
-
-  def apply[F[_]: ToFuture]: ToFuture[F] = summon[ToFuture[F]]
+object toFuture {
 
   given ioToFuture(using unsafe.IORuntime): ToFuture[IO] with {
-
     override def unsafeToFuture[A](fa: IO[A]): Future[A] = fa.unsafeToFuture()
-  }
-
-  given futureToFuture: ToFuture[Future] with {
-
-    override def unsafeToFuture[A](fa: Future[A]): Future[A] = fa
   }
 
   given idToFuture(using executionContext: ExecutionContext): ToFuture[Id] with {
