+++ cats-effect3/src/test/scala-2/effectie/cats/toFutureSpec.scala
@@ -1,0 +1,98 @@
+package effectie.cats
+
+import cats._
+import cats.effect._
+import effectie.cats.toFuture._
+import effectie.core.ToFuture
+import extras.concurrent.testing.ConcurrentSupport
+import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
+import hedgehog._
+import hedgehog.runner._
+
+import java.util.concurrent.ExecutorService
+import scala.concurrent.Future
+import scala.concurrent.duration._
+
+/** @author Kevin Lee
+  * @since 2020-09-23
+  */
+object toFutureSpec extends Properties {
+  private implicit val errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
+
+  override def tests: List[Test] = List(
+    property(
+      "test ToFuture[IO].unsafeToFuture",
+      IoSpec.testUnsafeToFuture
+    ),
+    property(
+      "test ToFuture[Id].unsafeToFuture",
+      IdSpec.testUnsafeToFuture
+    )
+  )
+
+  object IoSpec {
+
+    @SuppressWarnings(Array("org.wartremover.warts.IsInstanceOf"))
+    def testUnsafeToFuture: Property = for {
+      a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
+    } yield {
+      val expected = a
+      val fa       = IO(expected)
+
+      implicit val es: ExecutorService = ConcurrentSupport.newExecutorService(2)
+      @SuppressWarnings(Array("org.wartremover.warts.ExplicitImplicitTypes"))
+      implicit val ec                  =
+        ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
+      ConcurrentSupport.runAndShutdown(es, WaitFor(800.milliseconds)) {
+        import CatsEffectRunner._
+        implicit val ticket: Ticker = Ticker(TestContext())
+
+        val future   = ToFuture[IO].unsafeToFuture(fa)
+        val ioResult = fa.completeAs(expected)
+        val actual   = ConcurrentSupport.futureToValueAndTerminate(es, WaitFor(500.milliseconds))(future)
+
+        Result.all(
+          List(
+            ioResult,
+            Result
+              .assert(future.isInstanceOf[Future[Int]])
+              .log(s"future is not an instance of Future[Int]. future.getClass: ${future.getClass.toString}"),
+            actual ==== expected
+          )
+        )
+      }
+    }
+
+  }
+
+  object IdSpec {
+    @SuppressWarnings(Array("org.wartremover.warts.IsInstanceOf"))
+    def testUnsafeToFuture: Property = for {
+      a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
+    } yield {
+      implicit val es: ExecutorService = ConcurrentSupport.newExecutorService(2)
+
+      val fa          = a
+      @SuppressWarnings(Array("org.wartremover.warts.ExplicitImplicitTypes"))
+      implicit val ec =
+        ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
+      ConcurrentSupport.runAndShutdown(es, WaitFor(300.milliseconds)) {
+        val expected = Future(a)
+
+        val future = ToFuture[Id].unsafeToFuture(fa)
+        val actual = ConcurrentSupport.futureToValueAndTerminate(es, WaitFor(300.milliseconds))(future)
+
+        Result.all(
+          List(
+            Result
+              .assert(future.isInstanceOf[Future[Int]])
+              .log(s"future is not an instance of Future[Int]. future.getClass: ${future.getClass.toString}"),
+            actual ==== ConcurrentSupport.futureToValueAndTerminate(es, WaitFor(300.milliseconds))(expected),
+            actual ==== a
+          )
+        )
+      }
+    }
+  }
+
+}
