+++ cats-effect3/src/test/scala-3/effectie/cats/toFutureSpec.scala
@@ -2,32 +2,30 @@ package effectie.cats
 
 import cats.Id
 import cats.effect.*
-import effectie.cats.CatsEffectRunner.TestContext
 import cats.effect.unsafe.IORuntime
+import effectie.cats.CatsEffectRunner.*
 import effectie.cats.compat.CatsEffectIoCompatForFuture
+import effectie.cats.toFuture.given
+import effectie.core.ToFuture
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
 import hedgehog.*
 import hedgehog.runner.*
 
 import java.util.concurrent.ExecutorService
-import scala.concurrent.{ExecutionContext, Future}
 import scala.concurrent.duration.*
+import scala.concurrent.{ExecutionContext, Future}
 
 /** @author Kevin Lee
   * @since 2020-09-23
   */
-object ToFutureSpec extends Properties {
+object toFutureSpec extends Properties {
   private given errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
 
   override def tests: List[Test] = List(
     property(
       "test ToFuture[IO].unsafeToFuture",
       IoSpec.testUnsafeToFuture
-    ),
-    property(
-      "test ToFuture[Future].unsafeToFuture",
-      FutureSpec.testUnsafeToFuture
     ),
     property(
       "test ToFuture[Id].unsafeToFuture",
@@ -40,10 +38,11 @@ object ToFutureSpec extends Properties {
       a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
     } yield {
       val expected = a
-      val fa = IO(expected)
+      val fa       = IO(expected)
 
       given es: ExecutorService  = ConcurrentSupport.newExecutorService(2)
-      given ec: ExecutionContext = ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
+      given ec: ExecutionContext =
+        ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
       ConcurrentSupport.runAndShutdown(es, WaitFor(800.milliseconds)) {
         import effectie.cats.CatsEffectRunner.*
         given ticket: Ticker = Ticker(TestContext())
@@ -51,7 +50,7 @@ object ToFutureSpec extends Properties {
         val future   = ToFuture[IO].unsafeToFuture(fa)
         val ioResult = fa.completeAs(expected)
 
-        val actual   = ConcurrentSupport.futureToValueAndTerminate(es, WaitFor(500.milliseconds))(future)
+        val actual = ConcurrentSupport.futureToValueAndTerminate(es, WaitFor(500.milliseconds))(future)
 
         Result.all(
           List(
@@ -67,33 +66,6 @@ object ToFutureSpec extends Properties {
 
   }
 
-  object FutureSpec {
-    def testUnsafeToFuture: Property = for {
-      a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
-    } yield {
-      given es: ExecutorService  = ConcurrentSupport.newExecutorService(2)
-      given ec: ExecutionContext = ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
-      ConcurrentSupport.runAndShutdown(es, WaitFor(300.milliseconds)) {
-        val expected = Future(a)
-        val fa       = Future(a)
-
-        val future = ToFuture[Future].unsafeToFuture(fa)
-        val actual = ConcurrentSupport.futureToValueAndTerminate(es, WaitFor(300.milliseconds))(future)
-
-        Result.all(
-          List(
-            Result
-              .assert(future.isInstanceOf[Future[Int]])
-              .log(s"future is not an instance of Future[Int]. future.getClass: ${future.getClass.toString}"),
-            actual ==== ConcurrentSupport.futureToValueAndTerminate(es, WaitFor(300.milliseconds))(expected),
-            actual ==== a
-          )
-        )
-      }
-    }
-
-  }
-
   object IdSpec {
     def testUnsafeToFuture: Property = for {
       a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
@@ -101,7 +73,8 @@ object ToFutureSpec extends Properties {
       given es: ExecutorService = ConcurrentSupport.newExecutorService(2)
 
       val fa                     = a
-      given ec: ExecutionContext = ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
+      given ec: ExecutionContext =
+        ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
       ConcurrentSupport.runAndShutdown(es, WaitFor(300.milliseconds)) {
         val expected = Future(a)
