+++ core/src/test/scala-3/effectie/core/ToFutureSpec.scala
@@ -1,0 +1,53 @@
+package effectie.core
+
+import extras.concurrent.testing.ConcurrentSupport
+import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
+import hedgehog.*
+import hedgehog.runner.*
+
+import java.util.concurrent.ExecutorService
+import scala.concurrent.duration.*
+import scala.concurrent.{ExecutionContext, Future}
+
+/** @author Kevin Lee
+  * @since 2020-09-23
+  */
+object ToFutureSpec extends Properties {
+  private given errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
+
+  override def tests: List[Test] = List(
+    property(
+      "test ToFuture[Future].unsafeToFuture",
+      FutureSpec.testUnsafeToFuture
+    )
+  )
+
+  object FutureSpec {
+    def testUnsafeToFuture: Property = for {
+      a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
+    } yield {
+      given es: ExecutorService  = ConcurrentSupport.newExecutorService(2)
+      given ec: ExecutionContext =
+        ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
+      ConcurrentSupport.runAndShutdown(es, WaitFor(300.milliseconds)) {
+        val expected = Future(a)
+        val fa       = Future(a)
+
+        val future = ToFuture[Future].unsafeToFuture(fa)
+        val actual = ConcurrentSupport.futureToValueAndTerminate(es, WaitFor(300.milliseconds))(future)
+
+        Result.all(
+          List(
+            Result
+              .assert(future.isInstanceOf[Future[Int]])
+              .log(s"future is not an instance of Future[Int]. future.getClass: ${future.getClass.toString}"),
+            actual ==== ConcurrentSupport.futureToValueAndTerminate(es, WaitFor(300.milliseconds))(expected),
+            actual ==== a
+          )
+        )
+      }
+    }
+
+  }
+
+}
