+++ effectie-monix/src/main/scala/effectie/monix/toFuture.scala
@@ -2,6 +2,7 @@ package effectie.monix
 
 import cats.Id
 import cats.effect.IO
+import effectie.core.ToFuture
 import monix.eval.Task
 import monix.execution.Scheduler
 
@@ -10,15 +11,7 @@ import scala.concurrent.{ExecutionContext, Future}
 /** @author Kevin Lee
   * @since 2020-09-23
   */
-trait ToFuture[F[_]] {
-
-  def unsafeToFuture[A](fa: F[A]): Future[A]
-
-}
-
-object ToFuture {
-
-  def apply[F[_]: ToFuture]: ToFuture[F] = implicitly[ToFuture[F]]
+object toFuture {
 
   @SuppressWarnings(Array("org.wartremover.warts.ImplicitParameter"))
   implicit def taskToFuture(implicit scheduler: Scheduler): ToFuture[Task] = new ToFuture[Task] {
@@ -32,12 +25,6 @@ object ToFuture {
 
     override def unsafeToFuture[A](fa: IO[A]): Future[A] =
       fa.unsafeToFuture()
-  }
-
-  implicit val futureToFuture: ToFuture[Future] = new ToFuture[Future] {
-
-    override def unsafeToFuture[A](fa: Future[A]): Future[A] =
-      fa
   }
 
   @SuppressWarnings(Array("org.wartremover.warts.ImplicitParameter"))
